correction update

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -17,7 +17,7 @@
 void rev_string(char *s)
 {
 	long int slen = 0;
-	int x = 0;
+	long int x = 0;
 	char sc[10];
 
 	while (s[slen] != '\0')
